requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
django-channels==0.7.0
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
django-jazzmin==3.0.1
django-unfold==0.53.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
Faker==37.1.0
graphviz==0.20.3
idna==3.10
inflection==0.5.1
oauthlib==3.2.2
packaging==24.2
pillow==11.2.1
psycopg2==2.9.10
pydot==3.0.4
PyJWT==2.9.0
pyparsing==3.2.3
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
